{{cookiecutter.project_slug}}/requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
redis==3.3.11  # https://github.com/antirez/redis
requests==2.22.0
Werkzeug

# Django
# ------------------------------------------------------------------------------
django==2.2.6  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-admin-list-filter-dropdown
django-autocompletefilter
django-admin-shortcuts
django-cors-headers==3.0.2
django-cas-ng
-e git+https://gitlab+deploy-token-106917:ggxTmaGVQ8ZLSEPMMcKJ@gitlab.com/infoxel/tagx.git@master#egg=users&subdirectory=libs/users

{% if cookiecutter.db == 'django' %}
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
django-filter
{% endif %}


{% if cookiecutter.db == 'elastic' %}
elasticsearch-dsl>=7.0.0,<8.0.0
django-elasticsearch-dsl
{% endif %}

{% if cookiecutter.db == 'clickhouse' %}
infi.clickhouse_orm
{% endif %}
